requirements.txt
flask
flask-restful
flask-sqlalchemy
flask-migrate
flask-marshmallow
pymysql
python-dotenv
